disneys reservation website broken, kinda
tl;dr - not logged in the calendar is a complete lie from the
webpage (the robot is more legit), and logged in you must be
on a specific page for the calendar to refresh for real.

-----

when you are checking the MK calendar if you are not logged in,
the days are different from when you are logged in.

if you are not logged in, the calendar fetches dates randomly,
anywhere between once a minute and once an hour.

if you are logged in, you MUST refresh the page from this URL
https://disneyland.disney.go.com/entry-reservation/

if you do not go from this page and click the button, the
website does not actually ask the calendar for new reservation
data, it will instead feed you saved data. No matter who you
try to check or uncheck in your party, it doesn't matter. You
must return to that starting page URL and click the button or
the calendar won't refresh.

happy reserving.